logged out cart working

+++ browser/js/common/factories/CartService.js
@@ -1,7 +1,9 @@
-app.factory('CartService', function($state, $rootScope, $http, AuthService, Session, localStorageService){
+app.factory('CartService', function($state, $rootScope, $http, AuthService, Session, ProductService, localStorageService, $q){
 
 	var _cart = { line_items: [] };	
 	// _cart.line_items = localStorageService.get('cart');
+
+	//console.log('CartService.  Here is the LoggedInUser: ', LoggedInUser)
 
 
 	return {
@@ -16,13 +18,40 @@ app.factory('CartService', function($state, $rootScope, $http, AuthService, Sess
 			return _cart; 
 		},
 
-		findByUserId: function(userId, stateChange){
-			return $http.post('/api/users/' + userId + '/orders')
+		getLineItems: function(userId, stateChange){
+			// Logged-in user:
+			if (Session.user){
+				return $http.post('/api/users/' + userId + '/orders')
 				.then(function(result){
 					angular.copy(result.data, _cart);
 					$rootScope.$broadcast('cartChange', result.data, stateChange);
 					return _cart;
-				});
+				})
+			
+			// Unlogged user:	
+			} else {
+		        var tempArr = [];
+
+			    
+			    // Unlogged with something in local storage:
+			    if (localStorageService.get('cart').length  > 0) {
+	
+			    	_cart = { line_items: [] };
+
+			    	_cart.line_items = localStorageService.get('cart')
+
+    				$rootScope.$broadcast('cartChange', _cart);
+    				var deferred = $q.defer();
+			    	deferred.resolve(_cart);
+			    	return deferred.promise;
+			    } else {
+
+			    	//unlogged with nothing in local storage:
+			    	var deferred = $q.defer();
+			    	deferred.resolve(_cart);
+			    	return deferred.promise;
+			    }
+  			}
 		},
 
 		createOrder: function(userId){
@@ -34,20 +63,57 @@ app.factory('CartService', function($state, $rootScope, $http, AuthService, Sess
 		},
 
 		addLineItem: function(product){
-			var that = this;
-			return $http.post('/api/line_items/' + Session.user.id + '/order/' + _cart.id + '/line_items', { quantity: 1, productId: product.id} )
+			// Logged-in User:
+			if (Session.user){
+				var that = this;
+				return $http.post('/api/line_items/' + Session.user.id + '/order/' + _cart.id + '/line_items', { quantity: 1, productId: product.id} )
 				.then(function(result){
-					return that.findByUserId(Session.user.id)
+					return that.getLineItems(Session.user.id)
 				});
+			} else {
+
+				//Logged out user:
+
+			    	if (product){
+			    		// var that = this
+			    		var temp;
+			    		temp = localStorageService.get('cart');
+			    		temp.push({"product": product, "quantity": 1 , "id": -1});
+			    		
+			    		localStorageService.set('cart', temp);
+	
+
+						//clear out cart as getLIneItems should repopulate
+						_cart.line_items = localStorageService.get('cart');
+						var deferred = $q.defer();
+						deferred.resolve(_cart.line_items);
+						return deferred.promise;
+
+				}
+
+			}
+			
 		},
 
 		destroy: function(lineItem, index){
-			return $http.delete('/api/line_items/' + lineItem.id, { lineItemId: lineItem.id } )
-			.then(function(){
-				 _cart.line_items.splice(index,1);
-				 $rootScope.$broadcast('cartChange', _cart);
-				return _cart;	
-			});
+			if (Session.user){
+				return $http.delete('/api/line_items/' + lineItem.id, { lineItemId: lineItem.id } )
+				.then(function(){
+					_cart.line_items.splice(index,1);
+					console.log('_cart.line_items', _cart.line_items)
+					$rootScope.$broadcast('cartChange', _cart);
+					return _cart;	
+				});
+			} else {
+				_cart.line_items.splice(index,1);
+				console.log('_cart.line_items', _cart.line_items)
+ 				localStorageService.set('cart', _cart.line_items )
+				$rootScope.$broadcast('cartChange', _cart);
+				var deferred = $q.defer();
+				deferred.resolve(_cart);
+				return deferred.promise;
+
+			}	
 		},
 
 		update: function(lineItem){
